add team section
