+++ src/app/page.tsx
@@ -1,5 +1,6 @@
 // import { ExampleSection } from '@/views/ExampleSection';
 // import { ScrollToTopButton } from '@/components/ui/ScrollToTopButton';
+
 import { ExpertiseSection } from '@/views/ExpertiseSection';
 import { ApplyFormSection } from '@/views/ApplyFormSection';
 import { TimelineSection } from '@/views/TimelineSection';
