+++ src/data/footer.json
@@ -2,11 +2,11 @@
   "title": "Звʼяжися з нами",
   "terms": {
     "text": "Terms of use",
-    "href": ""
+    "href": "https://docs.google.com/document/d/1ilTzeNWTzwsxZQyvfB4o6nIg5WIHbvmM2FTuCmuRWz0/edit"
   },
   "policy": {
     "text": "Privacy Policy",
-    "href": ""
+    "href": "https://docs.google.com/document/d/1Ggeq7fAdERzW3lkE3RDQBjYi8MWkDEyGc8d51zOOz10/edit"
   },
   "email": {
     "text": "pentilosmm@gmail.com",
