+++ src/views/TimelineSection/TimelineSection.tsx
@@ -8,10 +8,7 @@ import { TimelineList } from '@/components/TimelineList';
 
 export const TimelineSection: React.FC = () => {
   return (
-    <section
-      className="section xl:h-[1072px] xl:pb-[39px] xl:pt-[58px]"
-      id={timeline}
-    >
+    <section className="section xl:h-[1130px]" id={timeline}>
       <div className="container xl:relative xl:flex xl:h-full">
         <h2 className="hidden">{data.hiddenTitle}</h2>
